Realm Object now compatible directly from Server.

+++ app/src/main/java/com/zemoso/atul/splitwise/adapters/FriendRecyclerViewAdapter.java
@@ -69,7 +69,7 @@ public class FriendRecyclerViewAdapter extends RecyclerView.Adapter<FriendRecycl
                 Toast.makeText(mContext,mCardData.getId() + " " + mCardData.getmHeading(),Toast.LENGTH_SHORT).show();
                 Intent mIntent = new Intent(mContext, FriendProfile.class);
                 Bundle mBundle = new Bundle();
-                mBundle.putInt("id",mCardData.getId());
+                mBundle.putLong("id", mCardData.getId());
                 mIntent.putExtras(mBundle);
                 mContext.startActivity(mIntent);
             }
